Add files via upload
This current version of my final project allows the user to interact with a GUI username and password manager. There are 2 text boxes: one for username and one for password. Then there are 5 buttons that the user can interact with. An Add button which adds the currently entered username and password combination into a txt list. Get button allows the user to enter a username and then it displays the password for that username. The List button displays all current usernames and passwords saved into the list. The Delete button allows the user to remove any saved username and password from the list. Then there is an Exit button at the top right which allows the user to exit the application.

I feel that I have a good foundation so far, but need to work on the visualization aspect of the application. I want to make it look better by adding colors, resizing some text and buttons, and then I still need to add some pictures to it.